requirements.txt
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
crispy-bootstrap5==0.7
decorator==5.1.1
Django==4.1.1
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-filter==22.1
executing==1.0.0
Faker==14.2.1
flake8==5.0.4
flake8-builtins==1.5.3
flake8-import-order==0.18.1
flake8-polyfill==1.0.2
ipython==8.5.0
jedi==0.18.1
matplotlib-inline==0.1.6
mccabe==0.7.0
parso==0.8.3
pep8-naming==0.13.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.13.0
python-dotenv==0.21.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.5.0
traitlets==5.4.0
wcwidth==0.2.5
webargs==8.2.0
